tools can be added/updated all via ajax. Modulated css/html/js

+++ modules/album/controllers/edit_album.php
@@ -20,8 +20,7 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
 	public function add($tool_id=NULL)
 	{
 		valid::id_key($tool_id);
-		echo $this->_view_add_single('album', $tool_id);
-		die();
+		die( $this->_view_add_single('album', $tool_id) );
 	}
 	
 /*
@@ -31,30 +30,31 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
  */
 	public function add_image($tool_id=NULL)
 	{
-		#TODO: TEST to make sure all names are unique. Cant have htem rewrite eachother.
+		# TODO: TEST to make sure all names are unique. 
+		# Cant have htem rewrite eachother.
 		valid::id_key($tool_id);
-		$db = new Database;	
-
-		// Work-around for setting up a session because Flash Player doesn't send the cookies
-		if (isset($_POST["PHPSESSID"])) {
-			session_id($_POST["PHPSESSID"]);
-		}
-		
+			
 		# validate for file size 5M and type (images)
-		if ( empty($_FILES['Filedata']['type']) OR ( $_FILES['Filedata']['type'] > 50000 ) )
-		{
+		if( empty($_FILES['Filedata']['type']) OR ( $_FILES['Filedata']['type'] > 50000 ) )
 			die('Invalid File');
-		}
+
+		# NOTE:: IS THIS SECURE??
+		# Work-around maintaining the session because Flash Player doesn't send the cookies
+		if(isset($_POST["PHPSESSID"]))
+			session_id($_POST["PHPSESSID"]);
+			
+		$db = new Database;	
 		
 		# Get highest position
 		$get_highest = $db->query("
 			SELECT MAX(position) as highest 
 			FROM album_items 
-			WHERE parent_id = '$tool_id' 
+			WHERE parent_id = '$tool_id'
+			AND fk_site = '$this->site_id'
 		")->current()->highest;
 
 		# Setup image store directory
-		$image_store = DOCROOT."data/$this->site_name/assets/images/albums";			
+		$image_store = DOCROOT . "data/$this->site_name/assets/images/albums";			
 		if(!is_dir($image_store))
 			mkdir($image_store);
 
@@ -86,9 +86,9 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
 			
 			# Make square thumbnails
 			if( $sm_width > $sm_height )
-				$image_sm->resize(120,120,Image::HEIGHT)->crop(120,120);
+				$image_sm->resize(100,100,Image::HEIGHT)->crop(100,100);
 			else
-				$image_sm->resize(120,120,Image::WIDTH)->crop(120,120);
+				$image_sm->resize(100,100,Image::WIDTH)->crop(100,100);
 			
 			$image_sm->save("$image_store/$tool_id/sm_$name");
 		}
@@ -126,10 +126,9 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
 	public function settings($tool_id=NULL)
 	{
 		valid::id_key($tool_id);		
-		$db = new Database;	
-	
 		if($_POST)
 		{
+			$db = new Database;	
 			$data = array(
 				'name'		=> $_POST['name'],
 				'view'		=> $_POST['view'],
@@ -151,21 +150,40 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
  * [see root JS in this::manage() ]
  * @PARM (INT) $id = id of image row 
  */
-	public function delete_image($id=NULL)
+	public function delete_image($id_string=NULL)
 	{
-		valid::id_key($id);		
-		
-		# Get image object
-		$image = $this->_grab_tool_child('album', $id);
+		$id_string = trim($id_string, '-');
+		$id_array = explode('-', $id_string);
+		foreach($id_array as $key => $id)
+			if(! is_numeric($id) )
+				unset($id_array[$key]);
 
-		# Image File delete
-		$image_path	= "$this->site_data_dir/assets/images/albums/$image->parent_id/$image->path";
-		$image_sm	= "$this->site_data_dir/assets/images/albums/$image->parent_id/sm_$image->path";
-		unlink($image_path);
-		unlink($image_sm);
+		if( '0' == count($id_array) )
+			die('invalid input');
+			
+		$id_string = implode(',',$id_array);
 		
-		$this->_delete_single_common('album', $id);
-		die();
+		foreach($id_array as $id)
+		{
+			# Get image object
+			$image = $this->_grab_tool_child('album', $id);
+
+			# Image File delete
+			$image_path	= "$this->site_data_dir/assets/images/albums/$image->parent_id/$image->path";
+			$image_sm	= "$this->site_data_dir/assets/images/albums/$image->parent_id/sm_$image->path";
+			if( file_exists($image_path) )
+				unlink($image_path);
+			if( file_exists($image_sm) )
+				unlink($image_sm);
+		}
+		
+		$db = new Database;
+		$db->query("
+			DELETE FROM album_items
+			WHERE id IN ($id_string)
+			AND fk_site = '$this->site_id'
+		");
+		die('images deleted');
 	}
 
 /*
@@ -175,9 +193,34 @@ class Edit_Album_Controller extends Edit_Tool_Controller {
  */
 	public function save_sort()
 	{
+		if( empty($_GET['image']) )
+			die('No items to sort');
+			
 		die( $this->_save_sort_common($_GET['image'], 'album_items') );
 	}
 	
+	static function _tool_adder($tool_id, $site_id)
+	{
+		return 'add';
+	}
 	
+/*
+ * delete all the images relating to this page and its folder
+ */
+	function _tool_deleter($tool_id, $site_id)
+	{
+		$album_dir = DATAPATH . "$this->site_name/assets/images/albums/$tool_id";
+		if(is_dir($album_dir))
+		{
+			$d = dir($album_dir); 
+			while($file = $d->read())
+			{
+				 if('.' != $file && '..' != $file)
+					unlink("$album_dir/$file"); 
+			} 
+			$d->close(); 
+			rmdir($album_dir);
+		}
+	}	
 }
 /* -- end of application/controllers/home.php -- */
